+++ openbank-dispute-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.9.2](https://github.com/JiRaska/open-bank-oss/compare/dispute-service-v0.9.1...dispute-service-v0.9.2) (2026-08-09)
+
+
+### Bug Fixes
+
+* **customer-edge:** answer 400, not 500, for a missing required query parameter ([#4211](https://github.com/JiRaska/open-bank-oss/issues/4211)) ([4ddb6ef](https://github.com/JiRaska/open-bank-oss/commit/4ddb6efeb23864fe65a4f2624f8722e1fcae04fb)), closes [#3624](https://github.com/JiRaska/open-bank-oss/issues/3624)
+
 ## [0.9.1](https://github.com/JiRaska/open-bank-oss/compare/dispute-service-v0.9.0...dispute-service-v0.9.1) (2026-08-08)